Move Method TrickList
From Home Controller to Trick Controller
